Make kActivityInitialProgress type `float`.

+++ components/ActivityIndicator/examples/supplemental/ActivityIndicatorExampleSupplemental.h
@@ -23,7 +23,7 @@
 #import "MaterialCollections.h"
 
 static const CGFloat kActivityIndicatorRadius = 72;
-static const CGFloat kActivityInitialProgress = (CGFloat)0.6;
+static const float kActivityInitialProgress = (float)0.6;
 
 @class ActivityIndicatorExample;
 @class MDCActivityIndicator;
